chore: prepare for adding auth
